+++ tests/test_serving.py
@@ -243,7 +243,7 @@ def test_https_redirect():
     app = AddOneModel(redirect_https=True)
     client = TestClient(app)
     add_one_response = client.get("/", allow_redirects=False)
-    assert add_one_response.status_code == 308
+    assert add_one_response.status_code == 307
 
 
 def test_predict_get_request():